LPS-75792 Move MonitorNames into internal package
