+++ modules/apps/foundation/portal/portal-monitoring/src/main/java/com/liferay/portal/monitoring/internal/statistics/portal/PortalRequestDataSample.java
@@ -15,8 +15,8 @@
 package com.liferay.portal.monitoring.internal.statistics.portal;
 
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.monitoring.MonitorNames;
 import com.liferay.portal.monitoring.internal.BaseDataSample;
+import com.liferay.portal.monitoring.internal.MonitorNames;
 
 /**
  * @author Rajesh Thiagarajan
